requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
binaryornot==0.4.4
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cryptography==44.0.1
daphne==4.1.2
Django==5.1.6
django-browser-reload==1.18.0
django-cotton==1.5.3
django-easy-audit==1.3.7
django-htmx==1.22.0
django-js-asset==3.1.2
django-mptt==0.16.0
django-tailwind==3.8.0
django-unfold==0.49.1
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zope.interface==7.2
